Refactored a bunch of the Analysis/AnalysisOutput code. I don't think the type parameters in those classes serve much of a purpose (assuming that all StreamingData objects are serializable), so I'm going to get rid of them
