+++ scala/src/main/scala/org/project/thunder/streaming/analyses/SimpleExampleAnalysis.scala
@@ -1,0 +1,8 @@
+package org.project.thunder.streaming.analyses
+
+/**
+ * Created by Andrew on 2/13/15.
+ */
+class SimpleExampleAnalysis {
+
+}
